CreatingAIEnabledSystems/LicensePlateRecognition/requirements.txt
numpy
opencv-python
ffmpeg-python
scikit-learn
matplotlib
pytesseract